[SR-4453] split AggregateOperator to sink-source pair
