+++ be/src/exec/pipeline/aggregate/aggregate_streaming_sink_operator.h
@@ -1,0 +1,49 @@
+// This file is licensed under the Elastic License 2.0. Copyright 2021 StarRocks Limited.
+
+#pragma once
+
+#include "exec/pipeline/operator.h"
+#include "exec/vectorized/aggregator.h"
+
+namespace starrocks::pipeline {
+class AggregateStreamingSinkOperator : public Operator {
+public:
+    AggregateStreamingSinkOperator(int32_t id, int32_t plan_node_id, AggregatorPtr aggregator)
+            : Operator(id, "aggregate_streaming_sink_operator", plan_node_id), _aggregator(aggregator) {}
+    ~AggregateStreamingSinkOperator() = default;
+
+    bool has_output() const override { return false; }
+    bool need_input() const override { return true; }
+    bool is_finished() const override;
+    void finish(RuntimeState* state) override;
+
+    Status prepare(RuntimeState* state) override;
+
+    StatusOr<vectorized::ChunkPtr> pull_chunk(RuntimeState* state) override;
+    Status push_chunk(RuntimeState* state, const vectorized::ChunkPtr& chunk) override;
+
+private:
+    Status _push_chunk_by_force_streaming();
+    Status _push_chunk_by_force_preaggregation(const size_t chunk_size);
+    Status _push_chunk_by_auto(const size_t chunk_size);
+
+    AggregatorPtr _aggregator;
+    // Whether prev operator has no output
+    bool _is_finished = false;
+};
+
+class AggregateStreamingSinkOperatorFactory final : public OperatorFactory {
+public:
+    AggregateStreamingSinkOperatorFactory(int32_t id, int32_t plan_node_id, AggregatorPtr aggregator)
+            : OperatorFactory(id, plan_node_id), _aggregator(aggregator) {}
+
+    ~AggregateStreamingSinkOperatorFactory() override = default;
+
+    OperatorPtr create(int32_t driver_instance_count, int32_t driver_sequence) override {
+        return std::make_shared<AggregateStreamingSinkOperator>(_id, _plan_node_id, _aggregator);
+    }
+
+private:
+    AggregatorPtr _aggregator;
+};
+} // namespace starrocks::pipeline
