--- be/src/exec/pipeline/aggregate_blocking_operator.h
@@ -1,35 +1,0 @@
-// This file is licensed under the Elastic License 2.0. Copyright 2021 StarRocks Limited.
-
-#pragma once
-
-#include "aggregate_base_operator.h"
-
-namespace starrocks::pipeline {
-class AggregateBlockingOperator final : public AggregateBaseOperator {
-public:
-    AggregateBlockingOperator(int32_t id, int32_t plan_node_id, const TPlanNode& tnode)
-            : AggregateBaseOperator(id, "aggregate_blocking", plan_node_id, tnode) {
-        _aggr_phase = vectorized::AggrPhase::AggrPhase2;
-    }
-    ~AggregateBlockingOperator() override = default;
-
-    bool has_output() const override;
-    bool is_finished() const override;
-    void finish(RuntimeState* state) override;
-
-    StatusOr<vectorized::ChunkPtr> pull_chunk(RuntimeState* state) override;
-    Status push_chunk(RuntimeState* state, const vectorized::ChunkPtr& chunk) override;
-};
-
-class AggregateBlockingOperatorFactory final : public AggregateBaseOperatorFactory {
-public:
-    AggregateBlockingOperatorFactory(int32_t id, int32_t plan_node_id, const TPlanNode& tnode)
-            : AggregateBaseOperatorFactory(id, plan_node_id, tnode) {}
-
-    ~AggregateBlockingOperatorFactory() override = default;
-
-    OperatorPtr create(int32_t driver_instance_count, int32_t driver_sequence) override {
-        return std::make_shared<AggregateBlockingOperator>(_id, _plan_node_id, _tnode);
-    }
-};
-} // namespace starrocks::pipeline
